// typescript/utils/src/types.ts
export enum ProtocolType {
  Ethereum = 'ethereum',
  Sealevel = 'sealevel',
  Cosmos = 'cosmos',
  CosmosNative = 'cosmosnative',
  Starknet = 'starknet',
  Radix = 'radix',
  Aleo = 'aleo',
  Tron = 'tron',
  Aeternity = 'aeternity',
  Unknown = 'unknown',
}